Analysis class (#7)
Created a class for managing an analysis of a record. Its should just for providing a simple api to this library, and chaining rule applications together in a nice way.

Its for sure not complete - but its a nice start. For one thing, the `analysis` property of an Analysis is just a big OrderedDict. Creating members of this ordered dict and processing it is messy and the code doesn't manage or control it. I have this intuition that some sort of tree structure for encapulating pieces and then doing things like bubbling up information from the lower levels of the Analysis would be better, but I haven't figured out how to do it yet.

* trying to debug a weird CRecord() initialization in tests with a case pre-loaded somehow in a new object

* fix: shouldn't instantiate an empty list as a default function arg, because python just creates the list once, and all instances of the function end up using the same exact list.

* fixing a bunch of little issues with handling missing values.