Refactor summary (#21)
* Fixed docker run command

* Fixed docker run command

* Refactor Summary Class, Part 1

This is the first stage of a refactoring of the Summary class.
The parse_pdf function now returns a Person (the defendant)
and a list of Cases.  These are stored in the Summary class
and returned by the get_defendant and get_cases mathods,
respectively.

In this refactor, there are three places where parse_pdf uses different
functions to process CP documents versus MD documents.  The function
parse_pdf as usual starts by calling pdftotext.  After that, it sends
the text to the function get_processors.  This function determines
which type of document is being processed and returns a dictionary
with the appropriate three processing functions.

Currently, get_processors checks the beginning of the text for the
word "Magisterial".  If we later find out that not all MD documents
have that word in their first line, we would have to change this function.

Currently, there is code duplication in the processing functions which
are in the RecordLib/summary.py file. The next step would be to
eliminate that.

This refactoring breaks tests which access summary._xml or summary.text.
I've changed those tests.  I also changed the fourth test in test_crecord.py.