more refactoring of Summary (#25)
* moved summary parsing to a directory of a few different submodules. Rewrote tests as well, to make each individual file a little simpler. And Summary takes a parser callable. The only callable now is a parse_pdf function, and Summary.__init__ uses that callable to read whatever raw source you give it.

* added more comments to utilities.py

* Another Refactor of the Summary Class (#26)

The Summary class is now a basic class which accepts a defendant and a list
of cases in its constructor.  The parse_pdf functions now returns a Summary
object.  The tests have been changed accordingly.

Going forward, the parse_pdf function and its accompanying functions would
be refactored into a class.