Django rest (#41)
* Fixed docker run command

* Fixed docker run command

* Add a Django REST backend

This is the start of a Django REST backend.
The backend is meant to be paired with a
React frontend.  The Django app is in the
directory "backend".   Currently, you can POST
a Summary pdf file to localhost:8000/upload.
The app will parse the file and return JSON with
the defendant's information and a list of the cases
from the file.