Add a React frontend (#45)
* Fixed docker run command

* Fixed docker run command

* Add a Django REST backend

This is the start of a Django REST backend.
The backend is meant to be paired with a
React frontend.  The Django app is in the
directory "backend".   Currently, you can POST
a Summary pdf file to localhost:8000/upload.
The app will parse the file and return JSON with
the defendant's information and a list of the cases
from the file.

* Add a React frontend

This is the start of a React frontend.
Currently, you can upload a Summary PDF.
The app will display information from the `CRecord`
generated from the Summary.

For instructions, see the React App section of the Readme.

* Edit React App instructions

* Add instructions for app development

* Add instructions and code fix for pdftotext

The Readme now has instructions for obtaining pdftotext.
Code has been edited to specify that the file produced by
pdftotext should be read with the uft8 encoding
(for operating systems for which this is not the default.)

* Remove unneeded import of pytest

pytest is a dev-package.

* Add Font Awesome icons

* Add and integrate a show/hide list component