Enable editing a case
Each case now has an Edit button, which changes its display
so that each field is in a textbox, and so that the user can add
charges and sentences.  The interface for adding a case is
integrated with this.  A new case is added to the end of the list
of cases and displayed in edit mode.  The user then enters
the data for the new case.