Bulk screen (#80)
Added to the analyze CLI - now we can analyze a whole bunch of files for obviously disqualifying issues. 

The `download_docs` cli can also use a table of urls to download documents. 

Created a couple ruledef functions for the triaging process, too.

This revealed that there are bugs w/ decision logic, but i think separate fixes are probably needed to fix these issues.

* added download_dockets subcommand to download dockets from a table of urls

* working on a cli  that looks to a directory of directories for records, and outputs a csv that states what records are ineligible based on certain rules.

* added a static method to Charge that tests if a charge's grade is greater than or equal to another, in seriousnes. i.e., a M1 is more serious than an S, but less serious than a F1.

* added a cli that triages a directory with lots of record pdfs in it. But lots of dockets still don't parse, and there are incorrect results, in particular false positivies, which are the worst for a triage system.

* tweaked format in a grammar, and added/removed breakpoints while debugging, but no substantive chagnes.