Combined record (#84)
* Fixed docker run command

* Fixed docker run command

* Change SentenceLength to a data class

SentenceLength is now a data class, with the generated
__init__ method.  The previous __init__ method has been
replaced with the class method from_tuples.

* Add key for alias list and fix typo

* Finish refactoring to have fewer connected components

* Add components for the applicant information

* Process CRecord to add aliases to the applicant

Also, the CRecord component now uses the Applicant component.

* Remove duplicate aliases from the list sent by the server

* Build a crecord from user input

* Add information from a summary pdf to an existing crecord

This commit also modifies the crecord state to make it simpler.
TODO: modify the code to return the crecord to the server so as to
take into account these changes to state.

* Update code to work with the recent changes to state

* Fix bugs in sending a crecord to be analyzed

* Fix bug and add feature in AddAttorney component

Before, pressing Enter while in the textbox triggered a form
submit.  Now, instead, it adds the attorney.