Add applicant name and date of birth from a CRecord
If the user has not already entered this information, the
information from an uploaded CRecord will be added to state and
displayed.  This commit also removes unused divs from the
RecordUploader component.  TODO, add instructions for the user to
this component.  Also, see if there is a way to inform the user
that their upload succeeded.