Applicant crecord (#91)
* Fixed docker run command

* Fixed docker run command

* Change SentenceLength to a data class

SentenceLength is now a data class, with the generated
__init__ method.  The previous __init__ method has been
replaced with the class method from_tuples.

* Add key for alias list and fix typo

* Finish refactoring to have fewer connected components

* Handle Enter key in AddCase and ServiceAgencyList

* Add applicant name and date of birth from a CRecord

If the user has not already entered this information, the
information from an uploaded CRecord will be added to state and
displayed.  This commit also removes unused divs from the
RecordUploader component.  TODO, add instructions for the user to
this component.  Also, see if there is a way to inform the user
that their upload succeeded.